+++ Basil/include/Core/Application.h
@@ -18,10 +18,22 @@ int main(int argc, char** argv);
 
 namespace Basil
 {
+	struct ApplicationCommandLineArgs
+	{
+		int count = 0;
+		char** args = nullptr;
+
+		const char* operator[](int index)
+		{
+			ASSERT(index < count, "");
+			return args[index];
+		}
+	};
+
 	class Application
 	{
 		public:
-			Application(const std::string& name = "Basil Application");
+			Application(const std::string& name = "Basil Application", ApplicationCommandLineArgs args = ApplicationCommandLineArgs());
 			virtual ~Application();
 			void close();
 			ImGuiLayer* getImGuiLayer();
@@ -29,11 +41,14 @@ namespace Basil
 			void pushLayer(Layer* layer);
 			void pushOverlay(Layer* layer);
 			static Application& get();
+			ApplicationCommandLineArgs getCommandLineArgs() const;
 			Window& getWindow();
 		private:
 			void run();
 			bool onWindowClose(WindowCloseEvent& e);
 			bool onWindowResize(WindowResizeEvent& e);
+			
+			ApplicationCommandLineArgs commandLineArgs;
 			Unique<Window> window;
 			ImGuiLayer* imGuiLayer;
 			bool running;
@@ -45,5 +60,5 @@ namespace Basil
 	};
 
 	// To be defined in the client
-	Application* createApplication();
+	Application* createApplication(ApplicationCommandLineArgs args);
 }