+++ premake5_yaml-cpp.lua
@@ -2,7 +2,6 @@ project "yaml-cpp"
 	location "Basil/vendor/"
 	kind "StaticLib"
 	language "C++"
-	staticruntime "On"
 	yamlcppdir = "Basil/vendor/yaml-cpp/"
 	targetdir ("bin/" .. outputdir .. "/%{prj.name}")
 	objdir ("bin-int/" .. outputdir .. "/%{prj.name}")